Optimizing samplings finder

+++ back/protected/controllers/CompanyController.php
@@ -53,16 +53,7 @@ class CompanyController extends Controller
 		);
 	}
 
-	/**a
-	 * Displays a particular model.
-	 * @param integer $id the ID of the model to be displayed
-	 */
-	public function actionView($id)
-	{
-		$this->render('view',array(
-			'model'=>$this->loadModel($id),
-		));
-	}
+	
 
 	/**
 	 * Creates a new model.
@@ -70,6 +61,8 @@ class CompanyController extends Controller
 	 */
 	public function actionCreate()
 	{
+    if(!Yii::app()->user->hasState('user'))
+      $this->redirect(array('site/login'));
 		$model=new Company;
 
 		// Uncomment the following line if AJAX validation is needed
@@ -92,6 +85,8 @@ class CompanyController extends Controller
 	
 	public function actionChange()
 	{
+    if(!Yii::app()->user->hasState('user'))
+      $this->redirect(array('site/login'));
 	  if(isset($_GET['company']))
 	  {
 	    $identity_company_model = IdentityCompany::model()->findByAttributes(array('company_id'=>$_GET['company'],'identity_id'=>Yii::app()->user->getState('user')));
@@ -113,103 +108,98 @@ class CompanyController extends Controller
 
   public function actionReset()
 	{
-	  if(Yii::app()->user->hasState('user'))
-	  {
-	    if(isset($_POST['reset_confirmation']))
-	    {
-	      
-	      $trucks = Truck::model()->findAllByAttributes(array('company_id'=>Yii::app()->user->getState('current_company')));
-        $trucks_ids = array();
-        foreach($trucks as $truck)
-          $trucks_ids[] = $truck->id;
-        
-        $criteria_find_routes = new CDbCriteria();
-        $criteria_find_routes->select = 'id';
-        $criteria_find_routes->addInCondition('truck_id', $trucks_ids);
-        $criteria_find_routes->with = 'beginning_stop';
-        $criteria_find_routes->with = 'end_stop';
-        
-        $routes = Route::model()->findAll($criteria_find_routes);
-	      $routes_ids = array();
-	      $long_stops_ids = array();
-	      foreach($routes as $route)
-	      {
-	        $routes_ids[] = $route->id;
-	        if(!empty($route->beginning_stop))
-	          $long_stops_ids[] = $route->beginning_stop->id;
-	        if(!empty($route->end_stop))
-	          $long_stops_ids[] = $route->end_stop->id;
-	      }
-	      
-	      //Sample
-	      $criteria_sample = new CDbCriteria();
-        $criteria_sample->addInCondition('truck_id', $trucks_ids);
-	      Sample::model()->deleteAll($criteria_sample);
-	      //sampling
-	      $criteria_sampling = new CDbCriteria();
-        $criteria_sampling->addInCondition('truck_id', $trucks_ids);
-	      Sampling::model()->deleteAll($criteria_sampling);
-	      //Shortstop
-	      $criteria_short_stop = new CDbCriteria();
-        $criteria_short_stop->addInCondition('route_id', $routes_ids);
-	      ShortStop::model()->deleteAll($criteria_short_stop);
-	      //route
-	      $criteria_route = new CDbCriteria();
-        $criteria_route->addInCondition('id', $routes_ids);
-	      Route::model()->deleteAll($criteria_route);
-	      //LongStop
-	      $criteria_long_stop = new CDbCriteria();
-        $criteria_long_stop->addInCondition('id', $long_stops_ids);
-	      LongStop::model()->deleteAll($criteria_long_stop);
-	      //Truck
-	      $criteria_truck = new CDbCriteria();
-        $criteria_truck->condition = 'company_id = '. Yii::app()->user->getState('current_company');
-	      Truck::model()->deleteAll($criteria_truck);
-	      
-	      $uploaded_files = UploadedFile::model()->findAllByAttributes(array('company_id'=>Yii::app()->user->getState('current_company')));
-	      foreach($uploaded_files as $uploaded_file)
-	      {
-	        unlink("../files/".$uploaded_file->filename);
-	        $uploaded_file->delete();
-	      }
-	      
-	      $company = Company::model()->findByPk(Yii::app()->user->getState('current_company'));
-	      $company->has_file_in_process = 0;
-	      $company->route_count = null;
-	      $company->average_speed = null;
-	      $company->average_stop_count_per_trip = null;
-	      $company->average_trip_distance = null;
-	      $company->average_stem_distance = null;
-	      $company->short_stop_time = null;
-	      $company->traveling_time = null;
-	      $company->resting_time = null;
-	      $company->time_radius_short_stop = null;
-	      $company->distance_radius_short_stop = null;
-	      $company->time_radius_long_stop = null;
-	      $company->distance_radius_long_stop = null;
-	      $company->distance_traveled = null;
-	      $company->average_short_stop_duration = null;
-	      $company->average_trip_duration = null;
-	      $company->average_trip_stop_time = null;
-	      $company->average_trip_traveling_time = null;
-	      $company->average_stop_count_per_trip_sd = null;
-	      $company->average_trip_distance_sd = null;
-	      $company->average_stem_distance_sd = null;
-	      $company->average_stem_distance_sd = null;
-	      $company->save();
-	      
-	      $this->render('successful_reset');
-	      
-	    }
-	    else
-	    {
-	      $this->render('reset_form');
-	    }
-	  }
-	  else
-	  {
-	    $this->redirect(array('site/login'));
-	  }
+    if(!Yii::app()->user->hasState('user'))
+      $this->redirect(array('site/login'));
+    if(isset($_POST['reset_confirmation']))
+    {
+      
+      $trucks = Truck::model()->findAllByAttributes(array('company_id'=>Yii::app()->user->getState('current_company')));
+      $trucks_ids = array();
+      foreach($trucks as $truck)
+        $trucks_ids[] = $truck->id;
+      
+      $criteria_find_routes = new CDbCriteria();
+      $criteria_find_routes->select = 'id';
+      $criteria_find_routes->addInCondition('truck_id', $trucks_ids);
+      $criteria_find_routes->with = 'beginning_stop';
+      $criteria_find_routes->with = 'end_stop';
+      
+      $routes = Route::model()->findAll($criteria_find_routes);
+      $routes_ids = array();
+      $long_stops_ids = array();
+      foreach($routes as $route)
+      {
+        $routes_ids[] = $route->id;
+        if(!empty($route->beginning_stop))
+          $long_stops_ids[] = $route->beginning_stop->id;
+        if(!empty($route->end_stop))
+          $long_stops_ids[] = $route->end_stop->id;
+      }
+      
+      //Sample
+      $criteria_sample = new CDbCriteria();
+      $criteria_sample->addInCondition('truck_id', $trucks_ids);
+      Sample::model()->deleteAll($criteria_sample);
+      //sampling
+      $criteria_sampling = new CDbCriteria();
+      $criteria_sampling->addInCondition('truck_id', $trucks_ids);
+      Sampling::model()->deleteAll($criteria_sampling);
+      //Shortstop
+      $criteria_short_stop = new CDbCriteria();
+      $criteria_short_stop->addInCondition('route_id', $routes_ids);
+      ShortStop::model()->deleteAll($criteria_short_stop);
+      //route
+      $criteria_route = new CDbCriteria();
+      $criteria_route->addInCondition('id', $routes_ids);
+      Route::model()->deleteAll($criteria_route);
+      //LongStop
+      $criteria_long_stop = new CDbCriteria();
+      $criteria_long_stop->addInCondition('id', $long_stops_ids);
+      LongStop::model()->deleteAll($criteria_long_stop);
+      //Truck
+      $criteria_truck = new CDbCriteria();
+      $criteria_truck->condition = 'company_id = '. Yii::app()->user->getState('current_company');
+      Truck::model()->deleteAll($criteria_truck);
+      
+      $uploaded_files = UploadedFile::model()->findAllByAttributes(array('company_id'=>Yii::app()->user->getState('current_company')));
+      foreach($uploaded_files as $uploaded_file)
+      {
+        unlink("../files/".$uploaded_file->filename);
+        $uploaded_file->delete();
+      }
+      
+      $company = Company::model()->findByPk(Yii::app()->user->getState('current_company'));
+      $company->has_file_in_process = 0;
+      $company->route_count = null;
+      $company->average_speed = null;
+      $company->average_stop_count_per_trip = null;
+      $company->average_trip_distance = null;
+      $company->average_stem_distance = null;
+      $company->short_stop_time = null;
+      $company->traveling_time = null;
+      $company->resting_time = null;
+      $company->time_radius_short_stop = null;
+      $company->distance_radius_short_stop = null;
+      $company->time_radius_long_stop = null;
+      $company->distance_radius_long_stop = null;
+      $company->distance_traveled = null;
+      $company->average_short_stop_duration = null;
+      $company->average_trip_duration = null;
+      $company->average_trip_stop_time = null;
+      $company->average_trip_traveling_time = null;
+      $company->average_stop_count_per_trip_sd = null;
+      $company->average_trip_distance_sd = null;
+      $company->average_stem_distance_sd = null;
+      $company->average_stem_distance_sd = null;
+      $company->save();
+      
+      $this->render('successful_reset');
+      
+    }
+    else
+    {
+      $this->render('reset_form');
+    }
 	}
 	
 	/**
@@ -219,6 +209,8 @@ class CompanyController extends Controller
 	 */
 	public function actionUpdate($id)
 	{
+    if(!Yii::app()->user->hasState('user'))
+      $this->redirect(array('site/login'));
 		$model=$this->loadModel($id);
 
 		// Uncomment the following line if AJAX validation is needed
@@ -245,22 +237,13 @@ class CompanyController extends Controller
 	 */
 	public function actionDelete($id)
 	{
+    if(!Yii::app()->user->hasState('user'))
+      $this->redirect(array('site/login'));
 		$this->loadModel($id)->delete();
 
 		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
 		if(!isset($_GET['ajax']))
 			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
-	}
-
-	/**
-	 * Lists all models.
-	 */
-	public function actionIndex()
-	{
-		$dataProvider=new CActiveDataProvider('Company');
-		$this->render('index',array(
-			'dataProvider'=>$dataProvider,
-		));
 	}
 
 	/**
